+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 
 * Tool retrieval is done with `use` instead of `get`
 * Tools are referenced by `SharedTool` (alias for `std::shared_ptr<ComputeTool>`)
+* Regions and profiles are subject to a translate-rotate-translate operation
+    - set by `origin`, `euler angles`, and `translation`
 
 ### Code Structure